+++ junit5-extension/src/main/java/io/kroxylicious/testing/kafka/junit5ext/Topic.java
@@ -3,10 +3,15 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.kroxylicious.testing.kafka.common;
+package io.kroxylicious.testing.kafka.junit5ext;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 
+/**
+ * Signals the wish for a topic to be created and injected into a test as either test parameter
+ * or field.
+ * {@see KafkaClusterExtension}
+ */
 public interface Topic {
     @NonNull
     String name();
