+++ junit5-extension/src/main/java/io/kroxylicious/testing/kafka/junit5ext/TopicPartitions.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.kroxylicious.testing.kafka.common;
+package io.kroxylicious.testing.kafka.junit5ext;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -12,9 +12,17 @@ import java.lang.annotation.Target;
 
 import io.kroxylicious.testing.kafka.api.KafkaClusterConstraint;
 
+/**
+ * Used to specify the number of partitions of a {@link Topic} injected into
+ * a test as field or parameter.
+ */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.FIELD, ElementType.PARAMETER })
 @KafkaClusterConstraint
 public @interface TopicPartitions {
+    /**
+     * The number of partitions.
+     * @return number of partitions.
+     */
     int value();
 }
