+++ junit5-extension/src/main/java/io/kroxylicious/testing/kafka/junit5ext/TopicReplicationFactor.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.kroxylicious.testing.kafka.common;
+package io.kroxylicious.testing.kafka.junit5ext;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -12,9 +12,18 @@ import java.lang.annotation.Target;
 
 import io.kroxylicious.testing.kafka.api.KafkaClusterConstraint;
 
+/**
+ * Used to specify the replication factor of a {@link Topic} injected into
+ * a test as field or parameter by the KafkaClusterExtension.
+ */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.FIELD, ElementType.PARAMETER })
 @KafkaClusterConstraint
 public @interface TopicReplicationFactor {
+    /**
+     * The topic replication factor.
+     *
+     * @return topic replication factor.
+     */
     short value();
 }
